Move base layer to separate package
